finish zen pomo record
