+++ src/components/Zen/ZenActions.vue
@@ -7,13 +7,13 @@
     </v-btn>
 
     <v-btn class='btn bg-secondary pomo-btn pomo-box font-press btn-main-start'
-      v-if="!pomodoro.going && !pomodoro.report?.shortPomo" @click="pomodoro.startPomodoro()">
+      v-if="!pomodoro.going && !pomodoro.finishedPomoRecord?.shortPomo" @click="pomodoro.startPomodoro()">
       <span>{{ $t("pause.study") }}</span>
       <v-icon class="icon" icon="mdi-play" />
     </v-btn>
 
     <v-btn class='btn bg-secondary pomo-btn pomo-box font-press btn-main-start'
-      v-if="!pomodoro.going && pomodoro.report?.shortPomo" @click="pomodoro.createPomodoro()">
+      v-if="!pomodoro.going && pomodoro.finishedPomoRecord?.shortPomo" @click="pomodoro.createPomodoro()">
       <span>{{ $t("backHome") }}</span>
       <v-icon class="icon" icon="mdi-home" />
     </v-btn>
@@ -34,7 +34,6 @@ const emit = defineEmits<{
   align-items: center;
   justify-content: center;
   flex-direction: row;
-  margin-top: 1em;
 }
 
 .btn {
